DR-main/requirements.txt
numpy
pandas
streamlit
tensorflow
pillow
keras
